requirements.txt
datasets==3.2.0
loguru==0.7.3
PyYAML==6.0.2
tokenizers==0.21.0
torch==2.3.0
transformers==4.47.1
typer==0.15.1
tqdm==4.66.4
wandb==0.17.4
sentencepiece==0.2.0
evaluate==0.4.3
sacrebleu==2.5.1
streamlit==1.41.1
fastapi==0.115.6
uvicorn==0.34.0
fastapi[standard]
